ege/requirements.txt
Django==1.11.3
django-markitup==3.0.0
docutils==0.13.1
Markdown==2.6.8
pytz==2017.2
